RNA_secondary_structure_prediction/requirements.txt
Django
matplotlib
networkx
numpy
pandas
python-igraph==0.7.1.post6
